Moved AggregatorGroup to its own file
